Transition to proper units tests for response class
